+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Added
+
+- Add support for getting nodepools.
+
 ### Removed
 
 - Remove the mention of the unexistent 'create cluster' command.